stable title - grok review - may 9 10:23
debug still doesn't get included in the changes window unless I change its name. not a huge deal, but it means I need to remember to do that each time for you to be able to assess those files... if there's an easy fix great, if not lets just stay focused on isolating columns in stable.py.